+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.21.0
  - Update to Flutter 3.22.0
  - feat: adding `NesDropdownMenu`
+ - feat: adding `NesContainer.borderColor` to allow customizing the border color.
 
 # 0.20.0
  - feat: adding different frames to `NesDialog`s